Add experimental overread support (lead-out)
This commit introduces experimental overread support for morituri. The
feature can be manually enabled through the command line switch `-x` or
`--force-overread`.

I’ve not tested it using a capable drive so I don’t know if it works as
expected.

The morituri log file should correctly report whether the rip was done
with the overread switch set or not (note that this does not imply that
the drive used is capable of overreading).

More informations are available through the integrated help.